Started working on Enemy AI
I made a enemy spaceship that follows the player while also dodging any objects that may be in the scene. It's good because it's dynamic, works only on raycasting, and doesn't need any baking to generate paths or anything. If it runs into something, it simply turns and continues trying to go to the player. Uses Quaternion.Slerp() to make it smoother